Fix: [AEA-5308] - Move the secrets and create a role to read them (#1219)
## Summary

- Routine Change

### Details

After a day or two trying to not do it this way, I've done it this way.

I need a role to access the secrets defined for the NHS notify tie-in.
This PR moves the secrets into the account_resources stack, and exports
a role that allows them to be read.